+++ src/ArduinoJson/Numbers/parseNumber.hpp
@@ -71,8 +71,9 @@ inline ParsedNumber<TFloat, TUInt> parseNumber(const char *s) {
       s++;
       break;
   }
-
+#if ARDUINOJSON_ENABLE_NAN
   if (*s == 'n' || *s == 'N') return traits::nan();
+#endif
   if (*s == 'i' || *s == 'I')
     return is_negative ? -traits::inf() : traits::inf();
   if (!isdigit(*s) && *s != '.') return return_type();